requirements.txt
firebase-admin==5.2.0
firebase==3.0.1
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
aws-xray-sdk==2.9.0
billiard==3.6.4.0
boto3==1.20.7
botocore==1.23.7
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coolsms-python-sdk==2.0.3
cryptography==35.0.0
daphne==3.0.2
Django==3.2.9
django-cors-headers==3.10.0
django-ckeditor==6.1.0
django-crontab==0.7.1
django-filter==21.1
django-hosts==4.0
django-import-export==2.6.1
django-safedelete==1.1.0
django-shortuuidfield==0.1.3
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
django-multiselectfield==0.1.12
django-nested-inline==0.4.4
django-ordered-model==3.4.3
drf-spectacular==0.21.0
drf-yasg==1.20.0
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-resources==5.4.0
incremental==21.3.0
inflection==0.5.1
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
reportlab==3.6.2
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0